move business logic into reusable module code
