+++ parsers/hml_parser.py
@@ -1,25 +1,7 @@
 """ HML parser! """
-
-# 3rd Party
-from twisted.internet import reactor
-from pyiem.nws.products.hml import parser as hmlparser
-
 # Local
-from pywwa import common
-from pywwa.ldm import bridge
-from pywwa.database import get_database
-
-
-def real_parser(txn, buf):
-    """I'm gonna do the heavy lifting here"""
-    prod = hmlparser(buf)
-    if common.dbwrite_enabled():
-        prod.sql(txn)
-    if prod.warnings:
-        common.email_error("\n".join(prod.warnings), buf)
+from pywwa.workflows.hml_parser import main
 
 
 if __name__ == "__main__":
-    bridge(real_parser, dbpool=get_database("hml"))
-
-    reactor.run()  # @UndefinedVariable
+    main()
