+++ tests/workflows/test_afos_dump.py
@@ -1,0 +1,14 @@
+"""Test afos_dump."""
+# 3rd Party
+import pytest
+
+# Local
+from pywwa.workflows import afos_dump
+from pywwa.testing import get_example_file, cursor
+
+
+@pytest.mark.parametrize("cursor", ["afos"])
+def test_processor(cursor):
+    """Test basic parsing."""
+    data = get_example_file("AFD.txt")
+    afos_dump.real_parser(cursor, data)
